Split ISS requests from CategoryPage
